refactor(theme): split into module + add Dracula + Tokyo Night
Reorganize theme code for maintainability:
  theme.rs → theme/mod.rs (struct defs + registry)
  theme/normal.rs, theme/cyberpunk.rs (existing, extracted)
  theme/dracula.rs (new — official Dracula palette)
  theme/tokyo_night.rs (new — official Tokyo Night palette)

Adding a new theme = one new file + one line in ALL_THEMES.

4 themes available: normal, cyberpunk, dracula, tokyo-night
Selectable via --theme flag or [t] in-app picker with live preview.
