+++ crates/ascii-agents/src/tui/theme/cyberpunk.rs
@@ -1,0 +1,91 @@
+use ascii_agents_core::sprite::Rgb;
+
+use super::*;
+
+pub static CYBERPUNK: Theme = Theme {
+    name: "cyberpunk",
+    surface: SurfaceColors {
+        wall: Rgb(22, 18, 35),
+        wall_trim: Rgb(50, 40, 70),
+        baseboard: Rgb(15, 12, 25),
+        carpet_base: Rgb(45, 42, 55),
+        carpet_light: Rgb(60, 55, 72),
+        carpet_dark: Rgb(32, 28, 42),
+        window_frame: Rgb(18, 14, 28),
+        bg_fallback: Rgb(12, 10, 20),
+    },
+    office: OfficeColors {
+        room_wall_body: Rgb(35, 28, 55),
+        room_wall_trim_light: Rgb(70, 55, 95),
+        room_wall_trim_dark: Rgb(20, 16, 32),
+        cubicle_divider: Rgb(50, 40, 75),
+        runner_base: Rgb(40, 35, 55),
+        runner_stripe: Rgb(60, 30, 80),
+        runner_edge: Rgb(25, 20, 38),
+        neon_panel_bg: Rgb(8, 6, 16),
+        neon_frame_base: Rgb(80, 20, 60),
+        building_dark: Rgb(12, 10, 22),
+        building_light: Rgb(30, 25, 50),
+        city_lit_window: Rgb(255, 60, 180),
+        city_dark_window: Rgb(18, 14, 30),
+        clock_rim: Rgb(120, 80, 200),
+        clock_face: Rgb(20, 15, 35),
+        clock_hand: Rgb(0, 255, 200),
+        shadow: Rgb(10, 8, 18),
+    },
+    lighting: LightingColors {
+        day_sky_a: Rgb(40, 20, 80),
+        day_sky_b: Rgb(60, 30, 100),
+        night_sky_a: Rgb(10, 6, 25),
+        night_sky_b: Rgb(20, 12, 45),
+        twilight_a: Rgb(180, 40, 120),
+        twilight_b: Rgb(220, 60, 160),
+        sun_spill: Rgb(200, 100, 255),
+        ceiling_pool: Rgb(120, 60, 255),
+        floor_lamp_halo: Rgb(0, 200, 255),
+        night_tint: Rgb(8, 6, 18),
+    },
+    furniture: FurnitureColors {
+        wood_top: Rgb(50, 45, 65),
+        wood_trim: Rgb(30, 25, 42),
+        rug_field: Rgb(40, 15, 60),
+        rug_trim: Rgb(25, 10, 38),
+        rug_accent: Rgb(150, 40, 120),
+        magazine: Rgb(60, 180, 255),
+        magazine_trim: Rgb(30, 90, 130),
+        chair_seat: Rgb(45, 40, 58),
+        chair_trim: Rgb(28, 24, 38),
+        coffee_cup: Rgb(80, 70, 100),
+        coffee_cup_shadow: Rgb(55, 48, 72),
+        desk_plant_light: Rgb(0, 255, 140),
+        desk_plant_dark: Rgb(0, 180, 100),
+        desk_plant_pot: Rgb(60, 50, 80),
+        photo_frame: Rgb(70, 50, 100),
+        photo_bg: Rgb(255, 60, 180),
+    },
+    effects: EffectColors {
+        monitor_frame_lit: Rgb(100, 60, 200),
+        sleep_z: Rgb(0, 200, 255),
+        coffee_steam: Rgb(0, 255, 140),
+        walking_dust: Rgb(60, 50, 80),
+        waiting_bubble: Rgb(255, 60, 180),
+    },
+    tool_glow: ToolGlowColors {
+        edit: Rgb(60, 120, 255),
+        read: Rgb(255, 60, 180),
+        bash: Rgb(255, 140, 0),
+        agent: Rgb(180, 0, 255),
+        grep: Rgb(0, 255, 140),
+        default: Rgb(0, 255, 200),
+    },
+    ui: UiColors {
+        label_active: Rgb(57, 255, 20),
+        label_waiting: Rgb(255, 60, 180),
+        label_idle: Rgb(80, 70, 120),
+        label_exiting: Rgb(40, 35, 60),
+        tooltip_bg: Rgb(10, 8, 20),
+        neon_brand: Rgb(255, 0, 200),
+        neon_star: Rgb(0, 255, 200),
+        neon_ticker: Rgb(120, 60, 255),
+    },
+};
