+++ crates/ascii-agents/src/tui/theme/dracula.rs
@@ -1,0 +1,91 @@
+use ascii_agents_core::sprite::Rgb;
+
+use super::*;
+
+pub static DRACULA: Theme = Theme {
+    name: "dracula",
+    surface: SurfaceColors {
+        wall: Rgb(40, 42, 54),
+        wall_trim: Rgb(68, 71, 90),
+        baseboard: Rgb(30, 31, 40),
+        carpet_base: Rgb(55, 52, 65),
+        carpet_light: Rgb(68, 64, 80),
+        carpet_dark: Rgb(42, 40, 52),
+        window_frame: Rgb(30, 31, 40),
+        bg_fallback: Rgb(40, 42, 54),
+    },
+    office: OfficeColors {
+        room_wall_body: Rgb(50, 52, 68),
+        room_wall_trim_light: Rgb(68, 71, 90),
+        room_wall_trim_dark: Rgb(30, 31, 40),
+        cubicle_divider: Rgb(68, 71, 90),
+        runner_base: Rgb(55, 50, 70),
+        runner_stripe: Rgb(68, 60, 85),
+        runner_edge: Rgb(40, 38, 52),
+        neon_panel_bg: Rgb(30, 31, 40),
+        neon_frame_base: Rgb(98, 114, 164),
+        building_dark: Rgb(25, 26, 35),
+        building_light: Rgb(50, 52, 68),
+        city_lit_window: Rgb(255, 184, 108),
+        city_dark_window: Rgb(30, 31, 40),
+        clock_rim: Rgb(189, 147, 249),
+        clock_face: Rgb(248, 248, 242),
+        clock_hand: Rgb(40, 42, 54),
+        shadow: Rgb(20, 20, 28),
+    },
+    lighting: LightingColors {
+        day_sky_a: Rgb(68, 71, 90),
+        day_sky_b: Rgb(98, 100, 120),
+        night_sky_a: Rgb(25, 26, 38),
+        night_sky_b: Rgb(35, 36, 52),
+        twilight_a: Rgb(189, 147, 249),
+        twilight_b: Rgb(255, 121, 198),
+        sun_spill: Rgb(255, 184, 108),
+        ceiling_pool: Rgb(248, 248, 242),
+        floor_lamp_halo: Rgb(255, 184, 108),
+        night_tint: Rgb(20, 20, 30),
+    },
+    furniture: FurnitureColors {
+        wood_top: Rgb(68, 71, 90),
+        wood_trim: Rgb(44, 45, 58),
+        rug_field: Rgb(60, 40, 70),
+        rug_trim: Rgb(40, 28, 50),
+        rug_accent: Rgb(189, 147, 249),
+        magazine: Rgb(139, 233, 253),
+        magazine_trim: Rgb(70, 116, 126),
+        chair_seat: Rgb(55, 56, 72),
+        chair_trim: Rgb(38, 39, 50),
+        coffee_cup: Rgb(130, 128, 145),
+        coffee_cup_shadow: Rgb(100, 98, 115),
+        desk_plant_light: Rgb(80, 250, 123),
+        desk_plant_dark: Rgb(50, 180, 80),
+        desk_plant_pot: Rgb(68, 60, 80),
+        photo_frame: Rgb(68, 71, 90),
+        photo_bg: Rgb(255, 121, 198),
+    },
+    effects: EffectColors {
+        monitor_frame_lit: Rgb(98, 114, 164),
+        sleep_z: Rgb(139, 233, 253),
+        coffee_steam: Rgb(189, 147, 249),
+        walking_dust: Rgb(68, 64, 80),
+        waiting_bubble: Rgb(241, 250, 140),
+    },
+    tool_glow: ToolGlowColors {
+        edit: Rgb(139, 233, 253),
+        read: Rgb(255, 121, 198),
+        bash: Rgb(255, 184, 108),
+        agent: Rgb(189, 147, 249),
+        grep: Rgb(80, 250, 123),
+        default: Rgb(139, 233, 253),
+    },
+    ui: UiColors {
+        label_active: Rgb(80, 250, 123),
+        label_waiting: Rgb(241, 250, 140),
+        label_idle: Rgb(98, 114, 164),
+        label_exiting: Rgb(68, 71, 90),
+        tooltip_bg: Rgb(30, 31, 40),
+        neon_brand: Rgb(189, 147, 249),
+        neon_star: Rgb(255, 121, 198),
+        neon_ticker: Rgb(139, 233, 253),
+    },
+};
