+++ crates/ascii-agents/src/tui/theme/normal.rs
@@ -1,0 +1,91 @@
+use ascii_agents_core::sprite::Rgb;
+
+use super::*;
+
+pub static NORMAL: Theme = Theme {
+    name: "normal",
+    surface: SurfaceColors {
+        wall: Rgb(56, 56, 70),
+        wall_trim: Rgb(80, 80, 100),
+        baseboard: Rgb(40, 40, 52),
+        carpet_base: Rgb(150, 110, 72),
+        carpet_light: Rgb(178, 138, 96),
+        carpet_dark: Rgb(118, 82, 50),
+        window_frame: Rgb(24, 24, 32),
+        bg_fallback: Rgb(28, 32, 40),
+    },
+    office: OfficeColors {
+        room_wall_body: Rgb(72, 74, 90),
+        room_wall_trim_light: Rgb(110, 112, 128),
+        room_wall_trim_dark: Rgb(40, 42, 54),
+        cubicle_divider: Rgb(72, 82, 104),
+        runner_base: Rgb(160, 120, 70),
+        runner_stripe: Rgb(140, 100, 55),
+        runner_edge: Rgb(90, 60, 35),
+        neon_panel_bg: Rgb(12, 14, 22),
+        neon_frame_base: Rgb(20, 60, 80),
+        building_dark: Rgb(20, 22, 32),
+        building_light: Rgb(60, 65, 82),
+        city_lit_window: Rgb(252, 215, 110),
+        city_dark_window: Rgb(30, 32, 44),
+        clock_rim: Rgb(200, 200, 210),
+        clock_face: Rgb(240, 240, 240),
+        clock_hand: Rgb(20, 20, 25),
+        shadow: Rgb(30, 25, 18),
+    },
+    lighting: LightingColors {
+        day_sky_a: Rgb(120, 160, 200),
+        day_sky_b: Rgb(160, 190, 220),
+        night_sky_a: Rgb(18, 26, 52),
+        night_sky_b: Rgb(28, 36, 70),
+        twilight_a: Rgb(220, 130, 80),
+        twilight_b: Rgb(240, 170, 110),
+        sun_spill: Rgb(255, 230, 160),
+        ceiling_pool: Rgb(255, 246, 215),
+        floor_lamp_halo: Rgb(255, 210, 130),
+        night_tint: Rgb(18, 22, 38),
+    },
+    furniture: FurnitureColors {
+        wood_top: Rgb(132, 88, 52),
+        wood_trim: Rgb(78, 52, 28),
+        rug_field: Rgb(140, 60, 50),
+        rug_trim: Rgb(90, 40, 35),
+        rug_accent: Rgb(190, 130, 80),
+        magazine: Rgb(98, 122, 178),
+        magazine_trim: Rgb(50, 60, 92),
+        chair_seat: Rgb(96, 68, 44),
+        chair_trim: Rgb(60, 40, 22),
+        coffee_cup: Rgb(200, 190, 170),
+        coffee_cup_shadow: Rgb(180, 160, 130),
+        desk_plant_light: Rgb(100, 180, 100),
+        desk_plant_dark: Rgb(60, 140, 60),
+        desk_plant_pot: Rgb(140, 100, 70),
+        photo_frame: Rgb(120, 100, 80),
+        photo_bg: Rgb(160, 200, 230),
+    },
+    effects: EffectColors {
+        monitor_frame_lit: Rgb(180, 200, 200),
+        sleep_z: Rgb(110, 110, 140),
+        coffee_steam: Rgb(190, 190, 210),
+        walking_dust: Rgb(150, 120, 85),
+        waiting_bubble: Rgb(255, 215, 70),
+    },
+    tool_glow: ToolGlowColors {
+        edit: Rgb(100, 160, 255),
+        read: Rgb(80, 220, 240),
+        bash: Rgb(240, 170, 80),
+        agent: Rgb(200, 140, 255),
+        grep: Rgb(180, 220, 120),
+        default: Rgb(140, 240, 170),
+    },
+    ui: UiColors {
+        label_active: Rgb(60, 220, 60),
+        label_waiting: Rgb(220, 200, 50),
+        label_idle: Rgb(140, 140, 140),
+        label_exiting: Rgb(80, 80, 80),
+        tooltip_bg: Rgb(20, 22, 30),
+        neon_brand: Rgb(80, 240, 255),
+        neon_star: Rgb(255, 100, 200),
+        neon_ticker: Rgb(180, 220, 255),
+    },
+};
