+++ crates/ascii-agents/src/tui/theme/tokyo_night.rs
@@ -1,0 +1,91 @@
+use ascii_agents_core::sprite::Rgb;
+
+use super::*;
+
+pub static TOKYO_NIGHT: Theme = Theme {
+    name: "tokyo-night",
+    surface: SurfaceColors {
+        wall: Rgb(26, 27, 38),
+        wall_trim: Rgb(65, 72, 104),
+        baseboard: Rgb(18, 18, 28),
+        carpet_base: Rgb(36, 40, 59),
+        carpet_light: Rgb(48, 52, 72),
+        carpet_dark: Rgb(26, 28, 42),
+        window_frame: Rgb(18, 18, 28),
+        bg_fallback: Rgb(26, 27, 38),
+    },
+    office: OfficeColors {
+        room_wall_body: Rgb(36, 40, 59),
+        room_wall_trim_light: Rgb(65, 72, 104),
+        room_wall_trim_dark: Rgb(22, 24, 36),
+        cubicle_divider: Rgb(65, 72, 104),
+        runner_base: Rgb(40, 44, 62),
+        runner_stripe: Rgb(52, 56, 78),
+        runner_edge: Rgb(28, 30, 44),
+        neon_panel_bg: Rgb(18, 18, 28),
+        neon_frame_base: Rgb(122, 162, 247),
+        building_dark: Rgb(18, 18, 28),
+        building_light: Rgb(36, 40, 59),
+        city_lit_window: Rgb(224, 175, 104),
+        city_dark_window: Rgb(22, 24, 36),
+        clock_rim: Rgb(169, 177, 214),
+        clock_face: Rgb(192, 202, 230),
+        clock_hand: Rgb(26, 27, 38),
+        shadow: Rgb(14, 14, 22),
+    },
+    lighting: LightingColors {
+        day_sky_a: Rgb(65, 72, 104),
+        day_sky_b: Rgb(90, 98, 130),
+        night_sky_a: Rgb(18, 18, 30),
+        night_sky_b: Rgb(26, 27, 42),
+        twilight_a: Rgb(187, 154, 247),
+        twilight_b: Rgb(247, 118, 142),
+        sun_spill: Rgb(224, 175, 104),
+        ceiling_pool: Rgb(192, 202, 230),
+        floor_lamp_halo: Rgb(224, 175, 104),
+        night_tint: Rgb(14, 14, 24),
+    },
+    furniture: FurnitureColors {
+        wood_top: Rgb(52, 56, 78),
+        wood_trim: Rgb(36, 38, 54),
+        rug_field: Rgb(45, 35, 65),
+        rug_trim: Rgb(30, 24, 45),
+        rug_accent: Rgb(187, 154, 247),
+        magazine: Rgb(122, 162, 247),
+        magazine_trim: Rgb(60, 80, 124),
+        chair_seat: Rgb(42, 46, 64),
+        chair_trim: Rgb(28, 30, 44),
+        coffee_cup: Rgb(110, 115, 140),
+        coffee_cup_shadow: Rgb(80, 84, 108),
+        desk_plant_light: Rgb(158, 206, 106),
+        desk_plant_dark: Rgb(100, 160, 70),
+        desk_plant_pot: Rgb(52, 50, 70),
+        photo_frame: Rgb(65, 72, 104),
+        photo_bg: Rgb(247, 118, 142),
+    },
+    effects: EffectColors {
+        monitor_frame_lit: Rgb(65, 72, 104),
+        sleep_z: Rgb(125, 207, 255),
+        coffee_steam: Rgb(187, 154, 247),
+        walking_dust: Rgb(48, 52, 72),
+        waiting_bubble: Rgb(224, 175, 104),
+    },
+    tool_glow: ToolGlowColors {
+        edit: Rgb(122, 162, 247),
+        read: Rgb(125, 207, 255),
+        bash: Rgb(224, 175, 104),
+        agent: Rgb(187, 154, 247),
+        grep: Rgb(158, 206, 106),
+        default: Rgb(125, 207, 255),
+    },
+    ui: UiColors {
+        label_active: Rgb(158, 206, 106),
+        label_waiting: Rgb(224, 175, 104),
+        label_idle: Rgb(65, 72, 104),
+        label_exiting: Rgb(45, 48, 65),
+        tooltip_bg: Rgb(18, 18, 28),
+        neon_brand: Rgb(122, 162, 247),
+        neon_star: Rgb(247, 118, 142),
+        neon_ticker: Rgb(125, 207, 255),
+    },
+};
